ENH: new method to split polydata in N contours actors
TODO: same for N contours polydatas
TODO: N time step doppler view
TODO: recompute volume/mesh too time consuming
TODO: 1 mesh = N xy, N xz, N yz contours + 1 3d actor?
TODO: 1 image = N xy, N xz, N yz images?
TODO: store meshes in ring buffer?

+++ Code/ExternalCode/MegaVTK/vtkRenderingAddOn/vtkViewImage2D.h
@@ -496,8 +496,8 @@ public:
       }
   }
 */
-  std::map<unsigned int, vtkActor *>
-  ExtractActors(vtkPolyData *iDataSet, ORIENTATION iOrientation);
+  std::map<double, vtkActor *> ExtractActors(
+    vtkPolyData *iDataSet, ORIENTATION iOrientation);
 
 protected:
